Move filters/index to wrapProvider
